requirements.txt
plotly==5.3.1
seaborn==0.11.2
streamlit==1.0.0
pandas==1.3.3
matplotlib==3.4.3
Pillow==8.3.2
